[rename] : member관련 파일 모두 member폴더로 이동
